C5
Tried to create a pixelated effect where the pixel were shapes whose corners' movements were generated with Perlin noise. Scrapped that idea because it was too complicated. Successfully finished another idea - a threshold black and white picture where the black area is the other picture. The threshold can be increased/decreased by the mouse, the pictures can be switched with the keyboard, and clicking switches the displayed section from black to white.